Don't let elements pending upgrade shadow properties when upgraded
lit-html extended allows using properties instead of attributes,
but when rending an element using custom elements which have not
been defined and upgraded yet, this will set properties which
when upgraded shadows over the lit-html-element property system.

This patch fixes that by checking for dependencies and postponing
the render.

This includes examples from Roger Urscheler, who pointed out this
issue.

Fixes: #22